==> 02-STseq Acute Wounds/s1-Seurat_STseq_Allsample.Rmd <==
---
title: "STseq essential analysis"
author: "Zhuang Liu"
date: "Updated: `r Sys.Date()`"
output: 
  html_document:
    toc: true #add a table of contents (TOC)
    toc_depth: 3 #headers' level in the toc
    toc_float: true #toc visible even when scrolled
    theme: lumen
    highlight: tango
    code_folding: show
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.align='center', fig.width=12, fig.height = 8, warning=FALSE, message=FALSE)
```

```{r warning=FALSE, message=FALSE}
library(Seurat)
library(tidyverse)
library(magrittr)
library(ggplot2)
library(patchwork)
library(sctransform)
library(cowplot)
library(harmony)
#library(future)
#plan("multisession", workers = 4)
#options(future.globals.maxSize = 20000 * 1024^2)
```

#####################################################
# Step 1: Load 10X filtered matrices into seurat
#####################################################
## 1.1 set up the Seurat objects
```{r}
samples <- list.files("/Users/zhuliu/Desktop/scRNA_STseq/proj_spatialTranseq/03_results/00-spacerange_counts/")
sm_folders <- paste0("/Users/zhuliu/Desktop/scRNA_STseq/proj_spatialTranseq/03_results/00-spacerange_counts/", samples, "/outs/")

stList <- list()
for (i in seq_along(samples)) {
  tmp_sms <- samples[i]
  stList[[tmp_sms]] <- Load10X_Spatial(sm_folders[i],
                                       filename = "filtered_feature_bc_matrix.h5", #default name
                                       assay = "Spatial") #default name
  stList[[tmp_sms]]$orig.ident <- tmp_sms
}

#Without any filtering, Check the spot numbers of each section
lapply(stList, ncol) %>% unlist()

tmp = stList[[1]]
SpatialFeaturePlot(tmp, features = "nFeature_Spatial")

plot(tmp@images$slice1@coordinates$row, tmp@images$slice1@coordinates$col)

x = seq(10, 70, 1)
y = -1.1 * x + 90
val = tmp@images$slice1@coordinates$col + 1.1 * tmp@images$slice1@coordinates$row
col = ifelse(val > 90, "red", "blue")
plot(tmp@images$slice1@coordinates$row, tmp@images$slice1@coordinates$col, col = col)
lines(x, y)

tmp = tmp[, val > 90]
SpatialFeaturePlot(tmp, features = "nFeature_Spatial")

stList[[1]] = tmp
```

```{r}
plots = list()
for (sample in names(stList)) {
    plots[[sample]] = SpatialFeaturePlot(stList[[sample]], features = "nFeature_Spatial")
}
wrap_plots(plots, ncol = 4)
```


## 1.2 Add the metadata
```{r}
metadata <- readxl::read_xlsx("../../02_metadata/Metadata_STsc.xlsx") %>% filter(Sequencing == "STseq")
metadata %<>% mutate(Sample_name = paste0(Donor, "_", Condition)) %>% select(2,4:7,12,15)

hswound.ST <- merge(stList[[1]], 
                    y = stList[2:length(stList)], 
                    add.cell.ids = metadata$Sample_name, 
                    project = "hsWoundST")

#Add the metadata
metadata <- hswound.ST@meta.data %>% rownames_to_column(var = "mapID") %>% left_join(., metadata, by=c("orig.ident" = "NGI_ID")) %>% column_to_rownames(var = "mapID")
hswound.ST <- AddMetaData(hswound.ST, metadata = metadata)

table(hswound.ST$Donor)
table(hswound.ST$Condition)

#rename images
names(hswound.ST@images) = unique(hswound.ST$Sample_name)
table(hswound.ST$Sample_name)
rm(stList);invisible(gc())
```


###########################
# Step 2: Quality control
###########################
## 2.1 Calculate QC metrics and draw Violin plots
```{r}
hswound.ST@active.ident = factor(hswound.ST$Sample_name)

#ratio of mitochondrial genes
hswound.ST[["percent.mt"]] <- PercentageFeatureSet(object = hswound.ST, pattern = "^MT-")
hswound.ST[["percent.ribo"]] <- PercentageFeatureSet(object = hswound.ST, pattern = "^RP[SL]")
#ratio of hemoglobin genes
hswound.ST[["percent.hb"]] <- PercentageFeatureSet(object = hswound.ST, pattern = "^HB[^(PES)]")
#ratio of MALAT1 genes
hswound.ST[["percent.malat1"]] <- PercentageFeatureSet(object = hswound.ST, pattern = "MALAT1")
```
```{r}
VlnPlot(hswound.ST, features = c("nFeature_Spatial", "nCount_Spatial"), ncol = 2, pt.size=0.0001)
VlnPlot(hswound.ST, features = c("percent.mt", "percent.ribo"), ncol = 2, pt.size=0.0001)
VlnPlot(hswound.ST, features = c("percent.hb", "percent.malat1"), ncol = 2, pt.size=0.0001)
```

## 2.2 FeaturePlots
```{r fig.height=8}
samples <- unique(hswound.ST$Sample_name)
for (i in samples) {
    plot_list <- SpatialFeaturePlot(hswound.ST[, hswound.ST$Sample_name == i], images = i, features = c("nFeature_Spatial", "nCount_Spatial", "percent.mt", "percent.ribo", "percent.hb", "percent.malat1"), ncol = 3)
    print(paste0("Ploting quality metrics of ", i))
    print(plot_list)
}
```


## 2.3 Inspect and filter the low QC spots
```{r}
mt.genes = grepl("^MT-", rownames(hswound.ST))
hswound.ST$nFeature_noMT = colSums(hswound.ST@assays$Spatial@counts[!mt.genes, ] > 0)

VlnPlot(hswound.ST, features = "nFeature_noMT")

dim(hswound.ST)

t1 = table(hswound.ST$Sample_name)

hswound.ST = hswound.ST[, hswound.ST$nFeature_noMT >= 100]

dim(hswound.ST)

t2 = table(hswound.ST$Sample_name)
t1 - t2
```

## 2.4 Top expressed gene
```{r}
####----Compute the top relative expression of each gene per spot Use sparse matrix----####
par(mar = c(4, 8, 2, 1))
C <- hswound.ST@assays$Spatial@counts
C <- Matrix::t(Matrix::t(C)/Matrix::colSums(C)) * 100
most_expressed <- order(apply(C, 1, median), decreasing = T)[20:1]
boxplot(t(as.matrix(C[most_expressed, ])), cex = 0.1, las = 1, xlab = "% total count per spot", 
    col = (scales::hue_pal())(20)[20:1], horizontal = TRUE)
```

```{r}
# Filter Mitocondrial
hswound.ST <- hswound.ST[!grepl("^MT-", rownames(hswound.ST)), ]

# Filter Hemoglobin gene (optional if that is a problem on your hswound.ST)
hswound.ST <- hswound.ST[!grepl("^HB[^(PES)]", rownames(hswound.ST)), ]

dim(hswound.ST)
```

## 2.5 Plotting after filtering
```{r fig.height=12}
VlnPlot(
  hswound.ST,
  features = c("nFeature_Spatial", "nCount_Spatial", "percent.mt", "percent.ribo", "percent.hb", "percent.malat1"),
  ncol = 2, pt.size=0)
```

## 2.6 FeaturePlots after filtering
```{r fig.height=8}
samples <- unique(hswound.ST$Sample_name)
for (i in samples) {
    plot_list <- SpatialFeaturePlot(hswound.ST[, hswound.ST$Sample_name == i], images = i, features = c("nFeature_Spatial", "nCount_Spatial", "percent.mt", "percent.ribo", "percent.hb", "percent.malat1"), ncol = 3)
    print(paste0("Ploting quality metrics of ", i))
    print(plot_list)
}
```


#######################################################
# Step 3: Preliminary clustering (no batch-correction)
#######################################################
## 3.1 Run sctransform, pca, clustering, umap
```{r}
hswound.ST <- SCTransform(hswound.ST, assay = "Spatial", verbose = FALSE)
hswound.ST <- RunPCA(hswound.ST, assay = "SCT", verbose = FALSE)
hswound.ST <- FindNeighbors(hswound.ST, reduction = "pca", dims = 1:30)
hswound.ST <- FindClusters(hswound.ST, resolution = 0.5)
hswound.ST <- RunUMAP(hswound.ST, reduction = "pca", dims = 1:30)
```

## 3.2 PCA plot
```{r fig.height=6, fig.width=8}
ElbowPlot(hswound.ST, ndims = 50, reduction = "pca")
```
```{r fig.height=18}
VizDimLoadings(hswound.ST, dims = 1:12, reduction = "pca")
```

## 3.3 Dim plot
```{r fig.height=8}
hswound.ST$Sample_name <- factor(hswound.ST$Sample_name, levels = unique(hswound.ST@meta.data$Sample_name))
p1 = DimPlot(hswound.ST, group.by = "Donor") + NoAxes()
p2 = DimPlot(hswound.ST, group.by = "Condition") + NoAxes()
sm.color <- cl.colors[1:16]
names(sm.color) <- unique(hswound.ST$Sample_name)

p3 = DimPlot(hswound.ST, group.by = "Sample_name", cols = sm.color) + NoAxes()
pdf("st_Fig_1_withoutHarmony.pdf", useDingbats = F, width = 12, height = 8)
wrap_plots(list(p1, p2, p3), ncol = 2)
dev.off()

DimPlot(hswound.ST, reduction = "umap", group.by = c("Sample_name", "Donor", "Condition"), ncol = 2)

SpatialDimPlot(hswound.ST) + plot_layout(ncol = 4)
```
```{r fig.height=5}
DimPlot(hswound.ST, reduction = "umap", group.by = "seurat_clusters", pt.size = .001, split.by = 'Donor', label = TRUE) + NoLegend() + ggtitle("Clusters by each donor")
```
```{r fig.height=5}
DimPlot(hswound.ST, reduction = "umap", group.by = "seurat_clusters", pt.size = .001, split.by = 'Condition', label = TRUE) + NoLegend() + ggtitle("Clusters by each condition")
```
```{r fig.height=5}
DimPlot(hswound.ST, reduction = "umap", group.by = "seurat_clusters", pt.size = .001, split.by = 'Seq_Batch', label = TRUE) + NoLegend() + ggtitle("Clusters by sequencing batch")
```

## 3.4 SpatialDim plot
```{r fig.height=12}
SpatialDimPlot(hswound.ST) + plot_layout(ncol = 4)
```

## 3.5 Plots for those metrics of quality
```{r}
# Determine metrics to plot present in hswound.STcom@meta.hswound.ST
metrics <- c("nCount_Spatial", "nFeature_Spatial", "percent.mt")

p3 <- FeaturePlot(
  hswound.ST,
  features = metrics,
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p3, function(x) {x + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 2)
VlnPlot(hswound.ST, features = metrics, group.by='seurat_clusters', pt.size = 0, ncol = 2)
```

## Integrate all
```{r}
hswound.ST@active.assay <- 'Spatial'
p = SpatialDimPlot(hswound.ST, combine = F)
wrap_plots(p, ncol = 4)

samples = unique(hswound.ST$Sample_name)
force = FALSE  #Set to true to rerun all calculations, otherwise will load saved files

# define color scale for clusters
cl.colors = c("#e6194B", "#3cb44b", "#ffe119", "#4363d8", "#f58231", "#911eb4", "#42d4f4",
    "#f032e6", "#bfef45", "#fabed4", "#469990", "#dcbeff", "#9A6324", "#fffac8",
    "#800000", "#aaffc3", "#808000", "#ffd8b1", "#000075", "#a9a9a9", "#000000")
names(cl.colors) = as.character(0:20)

ct.colors = c(RColorBrewer::brewer.pal(12, "Paired"), "cyan", "hotpink", "gray")

hswound.list <- SplitObject(hswound.ST, split.by = "Sample_name")
#Be careful that the image slots will contain 16 images, we need to filter them out
#hswound.list$Donor1_Skin@images
for (i in 1:length(names(hswound.list))) {
  print(names(hswound.list)[i])
  hswound.list[[i]]@images[setdiff(1:length(names(hswound.list)), i)] = NULL #remove unneeded images
  hswound.list[[i]] <- SCTransform(hswound.list[[i]], assay = "Spatial", verbose = FALSE)
  hswound.list[[i]] <- RunPCA(hswound.list[[i]], assay = "SCT")
}

vg = lapply(hswound.list, VariableFeatures)

# with all 3K genes
t = table(unlist(vg))
hist(t, main = "Number of datasets per variable gene (top 3K)")

vg.cond = list()
for (cond in unique(hswound.ST$Condition)) {
    m = grep(cond, names(vg))
    t = table(unlist(vg[m]))
    vg.cond[[cond]] = names(t)[t >= 3]
}

vg.all = unique(unlist(vg.cond))
length(vg.all)

# filter to only have genes detected in all sections.
rn = lapply(hswound.list, rownames)
nE = colSums(Reduce(rbind, lapply(rn, function(x) vg.all %in% x)))
vg.all = vg.all[nE == length(hswound.list)]
length(vg.all)

options(future.globals.maxSize = 15000 * 1024^2)  # set allowed size to 4K MiB
hswound.list <- PrepSCTIntegration(object.list = hswound.list, anchor.features = vg.all, verbose = FALSE)

data.merged <- merge(hswound.list[[1]], y = c(hswound.list[2:length(hswound.list)]), merge.data = TRUE)
data.merged$individual_clust = data.merged$seurat_clusters
```

## Batch correction
```{r}
VariableFeatures(data.merged) <- vg.all
gc(verbose = F)
data.merged <- RunPCA(object = data.merged, assay = "SCT", npcs = 40, verbose = F)
data.int <- RunHarmony(data.merged, assay.use = "SCT", reduction = "pca", dims.use = 1:40,
    group.by.vars = "Sample_name", plot_convergence = TRUE, verbose = T)

rm(data.merged);invisible(gc(verbose = F))
# run pca,umap,clustering.
data.int = RunUMAP(data.int, dims = 1:40, assay = "SCT", reduction = "harmony", spread = 1.5,
    min.dist = 0.1, verbose = F)
data.int = FindNeighbors(data.int, dims = 1:40, assay = "SCT", reduction = "harmony",
    verbose = F)
data.int = FindClusters(data.int, resolution = 0.5, verbose = F)

p1 = DimPlot(data.int, group.by = "Donor") + NoAxes()
p2 = DimPlot(data.int, group.by = "Condition") + NoAxes()
p3 = DimPlot(data.int, group.by = "seurat_clusters", label = T, cols = cl.colors) +
    NoAxes()
p4 = DimPlot(data.int, group.by = "Sample_name") + NoAxes()
wrap_plots(list(p1, p2, p3, p4), ncol = 2)

table(data.int$seurat_clusters)

plots = SpatialDimPlot(data.int, stroke = 0, combine = F, cols = cl.colors)
wrap_plots(plots, ncol = 4)

ggplot(data.int@meta.data, aes(x = seurat_clusters, fill = Donor)) + geom_bar(position = "fill")

ggplot(data.int@meta.data, aes(x = seurat_clusters, fill = Condition, alpha = Donor)) +
    geom_bar(position = "fill")
```

```{r eval=FALSE}
DefaultAssay(data.int) <- "Spatial" #DE for all genes instead of only using variable gene

data.int <- NormalizeData(data.int, verbose = TRUE,  normalization.method = "LogNormalize", scale.factor = 10000)

####----FindAllMarkers for all clusters----####
markers <- FindAllMarkers(data.int, only.pos = TRUE, assay = "Spatial") 

annot <- readRDS("./Functions/humanAnnotation.rds")
####----add the info and reorder the columns----####
markers_f <- markers %>% left_join(., annot, by=c("gene" = "external_gene_name")) %>% 
  dplyr::select(6, 7, 2:4, 1, 5, everything()) %>% arrange(cluster, desc(avg_log2FC))

variable.genes <- VariableFeatures(data.int, assay="SCT") %>% as.data.frame() %>% rename("gene" = ".") %>% 
  mutate(VariableGene = "Yes")
####----Add the information about if the gene is the variable gene----####
markers_f <- markers_f %>% left_join(., variable.genes, by=c("gene" = "gene"))
data.table::fwrite(markers_f, file="Seurat_STseq_integrated_markerGene.txt", sep="\t")
```


## Check some known cell type markers
### Well-known cell type markers
```{r eval=TRUE, fig.height=4, fig.width=12}
####----Marker genes----####
cc <- scale_color_gradientn(colors = c("grey", "yellow", "red", "black"))
small.leg <- theme(legend.text = element_text(size = 6), legend.key.width = unit(0.1, "cm"))

canonical_markers <- list(
  'Keratinocytes' = c('KRT5', 'KRT14', 'KRT10'), 
  'Fibroblasts' = c('COL1A1', 'COL1A2', 'LUM'), 
  'Melanocytes' =	c('PMEL',	'MLANA', 'TYRP1'), 
  'Endothelial-cells' = c('PECAM1', 'VWF', 'EGFL7'), 
  'Myeloid-cells' = c('LYZ', 'HLA-DRA', 'HLA-DRB1'), 
  'Mono-mac' = c('CD68', 'CD163', 'CD14'), #In dermis  
  'Macrophages' = c('CD209', 'FCGR2A', 'MRC1'), #CD209, FCGR2A aka CD32 or FCGR1A aka CD64, 
  #MRC1 aka CD206 (markers of macrophages, immature dendritic cells)
  'T-cells' = c('CD7', 'CD3D', 'CD3E'), 
  'T-cells2' = c('CD4', 'CD8A', 'NKG7'), 
  'B-cells' = c('MS4A1', 'CD79A', 'CD19'), 
  'Plasma-cells' = c('JCHAIN', 'IGHG1', 'MZB1'), 
  'LC-cells' = c('CD207', 'CD1A'), #Primarily located in epidermis (CD1A aka CD1)  
  'Dendritic-cells' = c('CD1C', 'CD1B', 'ITGAX'), #immature DC primarily located in superior dermis (ITGAX aka CD11C)
  'Dendritic-cells2' = c('CD83', 'LAMP3', 'THBD'), #mature DC: LAMP3 aka CD208, THBD aka CD141
  'Dendritic-cells3' = c('CLEC9A'), #CLEC9A aka DNGR1, a Dendritic Cell-Specific Sensor 
  'Schwann' = c('SOX10', 'SOX2'), 
  'PC-vSMC' = c('ACTA2', 'MCAM'), 
  'Epithelial_Mesenchymal' = c('EPCAM', 'PDGFRB')
)
plot_list <- FeaturePlot(
  data.int,
  features=unlist(canonical_markers),
  combine=FALSE
)


# apply theme to each feature plot
for(i in 1:length(plot_list)){
  plot_list[[i]] <- plot_list[[i]] + NoAxes() + cc + small.leg
}

cowplot::plot_grid(plotlist = plot_list[1:3], ncol = 3) + draw_figure_label(label = "Keratinocytes", size = 8, fontface = "bold", position = "top.right")
cowplot::plot_grid(plotlist = plot_list[4:6], ncol = 3) + draw_figure_label(label = "Fibroblasts", size = 8, fontface = "bold", position = "top.right")
cowplot::plot_grid(plotlist = plot_list[7:9], ncol = 3) + draw_figure_label(label = "Melanocytes", size = 8, fontface = "bold", position = "top.right")
cowplot::plot_grid(plotlist = plot_list[10:12], ncol = 3) + draw_figure_label(label = "Endothelial cells", size = 8, fontface = "bold", position = "top.right")
cowplot::plot_grid(plotlist = plot_list[13:15], ncol = 3) + draw_figure_label(label = "Myeloid Cells", size = 8, fontface = "bold", position = "top.right")
cowplot::plot_grid(plotlist = plot_list[16:18], ncol = 3) + draw_figure_label(label = "Monocytes & Macrophages", size = 8, fontface = "bold", position = "top.right")
cowplot::plot_grid(plotlist = plot_list[19:21], ncol = 3) + draw_figure_label(label = "Macrophages", size = 8, fontface = "bold", position = "top.right")
cowplot::plot_grid(plotlist = plot_list[22:24], ncol = 3) + draw_figure_label(label = "T/Th cells", size = 8, fontface = "bold", position = "top.right")
cowplot::plot_grid(plotlist = plot_list[25:27], ncol = 3) + draw_figure_label(label = "T/NK cells", size = 8, fontface = "bold", position = "top.right")
cowplot::plot_grid(plotlist = plot_list[28:30], ncol = 3) + draw_figure_label(label = "B cells", size = 8, fontface = "bold", position = "top.right")
cowplot::plot_grid(plotlist = plot_list[31:33], ncol = 3) + draw_figure_label(label = "Plasma cells", size = 8, fontface = "bold", position = "top.right")
cowplot::plot_grid(plotlist = plot_list[34:35], ncol = 3) + draw_figure_label(label = "Langerhans cells", size = 8, fontface = "bold", position = "top.right")
cowplot::plot_grid(plotlist = plot_list[36:38], ncol = 3) + draw_figure_label(label = "Dermal Dendritic cells", size = 8, fontface = "bold", position = "top.right")
cowplot::plot_grid(plotlist = plot_list[39:41], ncol = 3) + draw_figure_label(label = "Mature Dendritic cells", size = 8, fontface = "bold", position = "top.right")
cowplot::plot_grid(plotlist = plot_list[42], ncol = 3) + draw_figure_label(label = "Specific Dendritic cells", size = 8, fontface = "bold", position = "top.right")
cowplot::plot_grid(plotlist = plot_list[43:44], ncol = 3) + draw_figure_label(label = "Schwann cells", size = 8, fontface = "bold", position = "top.right")
cowplot::plot_grid(plotlist = plot_list[45:46], ncol = 3) + draw_figure_label(label = "PC-vSMC cells", size = 8, fontface = "bold", position = "top.right")
cowplot::plot_grid(plotlist = plot_list[47:48], ncol = 3) + draw_figure_label(label = "Epithelial / Mesenchymal cells", size = 8, fontface = "bold", position = "top.right")
```

```{r eval=FALSE}
saveRDS(data.int, file = "Seurat_STseq_integrated.rds")
```


# Sessioninfo
```{r}
sessionInfo()
```
