==> 02-STseq Acute Wounds/s3-STseq_integration_deconv.Rmd <==
---
title: "Stereoscope prediction"
author: "Zhuang Liu"
date: "Updated: `r Sys.Date()`"
output: 
  html_document:
    toc: true #add a table of contents (TOC)
    toc_depth: 3 #headers' level in the toc
    toc_float: true #toc visible even when scrolled
    theme: lumen
    highlight: tango
    df_print: paged
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.align='center', fig.width=12, fig.height = 7, warning=FALSE, message=FALSE)
```

```{r warning=FALSE, message=FALSE}
library(Seurat)
library(tidyverse)
library(magrittr)
library(ggplot2)
library(patchwork)
library(sctransform)
library(cowplot)
library(harmony)
require(gridExtra)
```

#####################################################
# Deconvolution results plotting
#####################################################
## SpatialDimPlot using Seurat
```{r fig.height=4}
rm(list = ls());invisible(gc())
# define color scale for clusters
cl.colors = c("#e6194B", "#3cb44b", "#ffe119", "#4363d8", "#f58231", "#911eb4", "#42d4f4",
    "#f032e6", "#bfef45", "#fabed4", "#469990", "#dcbeff", "#9A6324", "#fffac8",
    "#800000", "#aaffc3", "#808000", "#ffd8b1", "#000075", "#a9a9a9", "#000000")
names(cl.colors) = as.character(0:20)

hswound.STcom <- readRDS("Seurat_STseq_integrated.rds")
table(hswound.STcom$seurat_clusters)
p1 <- DimPlot(hswound.STcom, reduction = "umap", group.by='SCT_snn_res.0.5', label=T, cols = cl.colors) + NoAxes() + ggtitle('Cluster')
p2 <- DimPlot(hswound.STcom, reduction = "umap", group.by='Sample_name',) + NoAxes()
p1 + p2

hswound.STcom <- readRDS("/Users/zhuliu/Desktop/scRNA_STseq/proj_spatialTranseq/03_results/seurat_object_stereoscope_neut.rds")
rownames(hswound.STcom)
```

```{r fig.height=15}
img.names <- unique(hswound.STcom@meta.data$Sample_name) %>% as.character()

plot_list <- list()

for (i in seq_along(img.names)) {
  plot_list[i] <- SpatialDimPlot(hswound.STcom, group.by = "SCT_snn_res.0.5",images = img.names[i], stroke = 0.3, combine = F, pt.size.factor = 1.7, alpha = c(1, 1), image.alpha = 0.6, crop = TRUE, cols = cl.colors)  #, cols = cl.colors
}

# apply theme to each feature plot
for(i in 1:length(plot_list)) {
  plot_list[[i]] <- plot_list[[i]] + ggtitle(img.names[i])
}   

wrap_plots(plot_list, ncol = 4) + 
  plot_layout(guides = "collect") & theme(legend.position = 'bottom',
                                          legend.key.size = unit(1.2, 'cm'),
                                          legend.text = element_text(size=15),
                                          legend.title = element_text(size=15))
```

## Loading the deconvolution results
```{r}
#de_proport <- read_csv("/Users/zhuliu/Desktop/scRNA_STseq/proj_spatialTranseq/03_results/stereoscope_deconv/celltype_decon_proportions_Asa.csv")
de_proport <- read_csv("/Users/zhuliu/Desktop/scRNA_STseq/proj_spatialTranseq/03_results/stereoscope_deconv/uppmax_running/stereo_celltype_proportions.csv")
colnames(de_proport)[1] <- "barcode" 
#de_proport$barcode <- colnames(hswound.STcom)
identical(de_proport$barcode, colnames(hswound.STcom))
colnames(de_proport) <- gsub("-", "_", colnames(de_proport))
mt_decon <- hswound.STcom@meta.data %>% rownames_to_column(var = "barcode") %>% 
  left_join(., de_proport, by=c("barcode")) %>% column_to_rownames(var = "barcode")

hswound.STcom <- AddMetaData(hswound.STcom, metadata = mt_decon)
range(rowSums(de_proport[, -1]))
SpatialFeaturePlot(hswound.STcom, features = c('Basal_III', 'Granular_II'), images = "Donor2_Wound1", alpha = c(0.2, 1), combine = T)

# add as an assay object.
ste_count <- mt_decon %>% select(21:46) %>% t()
identical(colnames(ste_count), colnames(hswound.STcom))
celltypes=rownames(ste_count)
samples = sort(unique(hswound.STcom$Sample_name))

hswound.STcom[['Stereoscope']] <- CreateAssayObject(counts = ste_count)
```

### Plot functions
```{r}
plot.gene.umap = function(sdata, genes, reduction = "umap", assay = "Spatial", ncol=5, nrow=6, nPlot=NULL){
  cc = scale_color_gradientn(colors = c("grey","yellow","red","black"))
  small.leg <- theme(legend.text = element_text(size=6), legend.key.width = unit(0.1,"cm"))
  sdata@active.assay = assay
  pl = list()
  for (i in 1:length(genes)){
    pl[[i]] = FeaturePlot(sdata,features = genes[i], slot = "data",  pt.size = 0.01, order = T, reduction = reduction) + NoAxes() + cc + small.leg + ggtitle(genes[i]) + theme(plot.title = element_text(size=10))
  }
  if (is.null(nPlot)){ nPlot = length(genes)}
  if (nPlot > ncol*nrow){
    print("Cannot fit all figures into a single plot")
    nPlot = ncol*nrow
  }
  nBin = ceiling(length(pl)/nPlot)
  for (i in 1:nBin){
    x = (i*nPlot+1-nPlot):min(length(pl),i*nPlot)
    grid.arrange(grobs=pl[x], ncol=ncol, nrow=nrow)
  }
}
```
```{r}
plot.gene.st = function(sdata,genes,images,assay="Spatial", ncol=5, nrow=6, nPlot=NULL){
  
  tmp = data.frame(t(sdata@assays[[assay]]@counts[genes,]))
  tmp$x = sdata@images[[sample]]@coordinates$imagecol
  tmp$y = -sdata@images[[sample]]@coordinates$imagerow
  
  cc = scale_color_gradientn(colors = c("gray","yellow","red","black"))
  small.leg <- theme(legend.text = element_text(size=6), legend.key.width = unit(0.1,"cm"))
  pl = list()
  for (g in genes){
    g2 = sym(g)
    pl[[g]] = ggplot(tmp, aes(x=x,y=y,color= !! g2)) + geom_point()  + theme_classic() + cc + ggtitle(g) + NoAxes() + small.leg
  }
  if (is.null(nPlot)){ nPlot = length(genes)}
  if (nPlot > ncol*nrow){
    print("Cannot fit all figures into a single plot")
    nPlot = ncol*nrow()
  }
  nBin = ceiling(length(pl)/nPlot)
  for (i in 1:nBin){
    x = (i*nPlot+1-nPlot):min(length(pl),i*nPlot)
    grid.arrange(grobs=pl[x], ncol=ncol, nrow=nrow)
  }
}
```

## Stereoscope results projected to umap of clusters in ST data
```{r, fig.height=25}
celltypes <- rownames(hswound.STcom)
plot.gene.umap(hswound.STcom, celltypes, assay = "Stereoscope", ncol = 4, nrow = 7)
```

## Stereoscope results projected to features of HE staining slides in ST data
```{r, fig.width=12, fig.height=12}
hswound.STcom@active.assay = "Stereoscope"

for (g in celltypes){
  p = SpatialFeaturePlot(hswound.STcom, features = g, images = samples, combine = F, alpha = c(0.2, 1))
  grid.arrange(grobs=p, ncol = 4, top = g)
}
```


#####################################################
# Broad celltype deconvolution results plotting
#####################################################
## Add up total cell type proportions by 4 main celltypes
```{r}
SpatialFeaturePlot(hswound.STcom, features = c('Basal_III', 'Granular_II'), images = "Donor2_Wound1", combine = T)
```
```{r eval=True}
anno <- readxl::read_xlsx("/Users/zhuliu/Desktop/scRNA_STseq/proj_10X_woundhealing/03_results/01-Seurat-PreAnalysis/02_Seurat_BatchCorrection/s1_Seurat_allSample_harmony/s5_cell_types_for_each_cluster_manually_updated.xlsx")
anno = anno[!(anno$CellType == "Basal-IV"), ]
anno$CellType = gsub("_", "-", anno$CellType)

anno$Group = anno$MainCellTypes

anno$Group[anno$Group %in% c("PlasmaBcells", "NKTcells", "MyeloidCells", "MastCells",
    "LangerhansCells")] = "Immune"
anno$Group[anno$Group %in% c("EndothelialCells", "PericytesSmoothMuscleCells",
    "SchwannCells", "Melanocytes")] = "Other"

ct.groups = split(anno$CellType, anno$Group)
```

########################
# Celltype proportion
########################
## Per section
```{r}
samples <- unique(hswound.STcom$Sample_name)
cell.prop = list()
for (s in samples) {
    cell.propDW = rowSums(hswound.STcom@assays$Stereoscope@counts[, hswound.STcom$Sample_name ==
        s])
    cell.propDW = cell.propDW/sum(cell.propDW)
    cell.prop[[s]] = cell.propDW
}

cell.prop = Reduce(cbind, cell.prop)
colnames(cell.prop) = samples

tot.prop = data.frame(cell.prop)
# reoder to have all Skin next to each other etc.
s = unlist(lapply(strsplit(samples, "_"), function(x) x[2]))
o = order(s)
tot.prop = tot.prop[, o]

tot.prop$celltype = rownames(tot.prop)

l = reshape2::melt(tot.prop, id.vars = "celltype")
l$celltype = factor(l$celltype, levels = anno$CellType)

# define the colors
anno = anno[order(anno$MainCellTypes, anno$CellType), ]
anno$Color = NA
anno[grepl("Basal", anno$CellType), "Color"] = colorRampPalette(c("green", "darkgreen"))(3)
anno[grepl("Spinous", anno$CellType), "Color"] = colorRampPalette(c("lightblue",
    "darkblue"))(3)
anno[grepl("Granular", anno$CellType), "Color"] = c("cyan", "cyan3")
anno[grepl("FB", anno$CellType), "Color"] = colorRampPalette(c("pink", "purple3"))(4)
anno[anno$Group == "Other", "Color"] = colorRampPalette(c("grey82", "gray30"))(5)
anno[is.na(anno$Color), "Color"] = colorRampPalette(c("yellow", "red"))(9)

ct.cols = anno$Color
names(ct.cols) = anno$CellType

p5 = ggplot(l, aes(x = variable, y = value, fill = celltype)) + geom_bar(position = "fill",
    stat = "identity") + RotatedAxis() + scale_fill_manual(values = ct.cols)
print(p5)
```

```{r, fig.height=10}
# (Violin plots) Plot 9 at a time
p = VlnPlot(hswound.STcom, features = celltypes,  group.by = "Sample_name", pt.size = 0, combine = F)
p = lapply(p, function(x) x + theme(axis.text.x = element_text(size = 6)) + NoLegend())


grid.arrange(grobs = p[1:9], ncol=3)
grid.arrange(grobs = p[10:18], ncol=3)
grid.arrange(grobs = p[19:26], ncol=3)
```

## Per cluster
```{r}
cell.prop.cl = list()
for (cl in unique(hswound.STcom$seurat_clusters)) {
    cell.propDW = rowSums(hswound.STcom@assays$Stereoscope@counts[, hswound.STcom$seurat_clusters ==
        cl])
    cell.propDW = cell.propDW/sum(cell.propDW)
    cell.prop.cl[[cl]] = cell.propDW
}

cell.prop.cl = Reduce(cbind, cell.prop.cl)
colnames(cell.prop.cl) = unique(hswound.STcom$seurat_clusters)

cell.prop.cl = cell.prop.cl[, order(as.numeric(colnames(cell.prop.cl)))]

tot.prop = data.frame(cell.prop.cl)

tot.prop$celltype = rownames(tot.prop)

l = reshape2::melt(tot.prop, id.vars = "celltype")
l$celltype = factor(l$celltype, levels = anno$CellType)

p5 = ggplot(l, aes(x = variable, y = value, fill = celltype)) + geom_bar(position = "fill",
    stat = "identity") + RotatedAxis()
p5 + scale_fill_manual(values = ct.cols)
```




```{r eval=TRUE}
canonical_markers <- list(
  'Keratinocytes' = c('KRT5', 'KRT14', 'KRT10'), 
  'Fibroblasts' = c('COL1A1', 'COL1A2', 'LUM'), 
  'Melanocytes' =	c('PMEL',	'MLANA', 'TYRP1')
)

genes <- unlist(canonical_markers)

plot_list <- list()
for (i in seq_along(genes)) {
  plot_list[[i]] <- SpatialFeaturePlot(hswound.STcom, features = genes[i],images = "Donor1_Skin", stroke = 0.4, combine = T, pt.size.factor = 1.7, alpha = c(1, 1), image.alpha = 0.6, crop = TRUE)
}

patchwork::wrap_plots(plot_list[1:3], ncol = 3) + plot_annotation(title = "Keratinocytes")
```


# Sessioninfo
```{r}
sessionInfo()
```
